put request added to the list

+++ facite/lists/api/views.py
@@ -57,15 +57,35 @@ class ListDetailAPIView(_APIView):
     def get(self, request, list_id):
         user = request.user
         list_obj = _get_object_or_404(self.model, pk=list_id)
+        if list_obj.user != user:
+            return _response.Response(status=_status.HTTP_404_NOT_FOUND)
 
         serializer = _ListSerializer(list_obj)
 
         return _response.Response(data=serializer.data)
 
+    @_transaction.atomic
     def delete(self, request, list_id):
         user = request.user
         list_obj = _get_object_or_404(self.model, pk=list_id)
+        if list_obj.user != user:
+            return _response.Response(status=_status.HTTP_404_NOT_FOUND)
 
         list_obj.delete()
 
         return _response.Response(status=_status.HTTP_204_NO_CONTENT)
+
+    @_transaction.atomic
+    def put(self, request, list_id):
+        user = request.user
+        serializer = _ListSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        list_obj = _get_object_or_404(self.model, pk=list_id)
+        list_obj.name = serializer.validated_data["name"]
+        list_obj.is_completed = serializer.validated_data["is_completed"]
+        list_obj.date_completed = serializer.validated_data["date_completed"]
+        list_obj.save()
+
+        serializer = _ListSerializer(instance=list_obj)
+
+        return _response.Response(serializer.data)
